requirements.txt
bidict==0.21.4
Flask==2.0.3
Flask_SocketIO==5.1.1
imgviz==1.4.2
matplotlib==3.5.1
numpy==1.22.2
opencv_python==4.5.5.62
pandas==1.5.0
Pillow==9.4.0
python-socketio==5.10.0
PyYAML==6.0
rsa==4.9
scikit_learn==1.1.2
seaborn==0.13.0
segmentation_models_pytorch==0.3.0
tensorboardX==2.6
timm==0.4.12
torchkeras==3.2.5
tqdm==4.62.3
Werkzeug==2.2.2
scipy
accelerate
